feat: restructuring, exporting, documentation
