+++ packages/vc-export/src/index.ts
@@ -2,5 +2,12 @@ import verification from './verificationUtils'
 import presentation from './presentationUtils'
 import { fromAttestedClaim } from './exportToVerifiableCredential'
 import * as types from './types'
+import * as vcjsSuites from './vc-js'
 
-export default { verification, presentation, fromAttestedClaim, types }
+export default {
+  verification,
+  presentation,
+  fromAttestedClaim,
+  types,
+  vcjsSuites,
+}
