+++ packages/vc-export/src/vc-js/documentLoader.ts
@@ -1,6 +1,6 @@
 import { RemoteDocument, Url } from 'jsonld/jsonld-spec'
 import vcjs from 'vc-js'
-import kiltContexts from './kiltContexts'
+import kiltContexts from './context'
 
 export default async function documentLoader(
   url: Url
