+++ packages/vc-export/src/vc-js/examples/example-vc.json
