+++ packages/vc-export/src/vc-js/index.ts
@@ -1,0 +1,5 @@
+import * as suites from './suites'
+import documentLoader from './documentLoader'
+import context, { KILT_CREDENTIAL_CONTEXT_URL } from './context'
+
+export { suites, documentLoader, context, KILT_CREDENTIAL_CONTEXT_URL }
