+++ packages/vc-export/src/vc-js/suites/index.ts
@@ -1,0 +1,5 @@
+import KiltIntegritySuite from './KiltIntegritySuite'
+import KiltSignatureSuite from './KiltSignatureSuite'
+import KiltAttestedSuite from './KiltAttestedSuite'
+
+export { KiltIntegritySuite, KiltSignatureSuite, KiltAttestedSuite }
